Rework LineageOS unfolding process
Unlike AOSP, LineageOS does not have static tags.
Therefore current fetching approach does not work as intended.

To solve the issue, we now store a copy of the static manifest
file for LineageOS in a repository.

To update the LineageOS static manifest, the maintainer shall use
the update_lineage_manifest.sh script.

Signed-off-by: Roman Stratiienko <r.stratiienko@gmail.com>